procurement_neg_agents/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
authlib==1.6.0
    # via google-adk
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   google-adk
    #   uvicorn
colorama==0.4.6
    # via click
cryptography==45.0.3
    # via authlib
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
docstring-parser==0.16
    # via google-cloud-aiplatform
exceptiongroup==1.3.0
    # via anyio
fastapi==0.115.12
    # via google-adk
google-adk==1.1.1
    # via -r .\requirements.in
google-api-core[grpc]==2.25.0rc1
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
google-api-python-client==2.170.0
    # via google-adk
google-auth==2.40.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform==1.95.1
    # via google-adk
google-cloud-bigquery==3.34.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.21.0
    # via -r .\requirements.in
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.23.3
    # via google-adk
google-cloud-speech==2.32.0
    # via google-adk
google-cloud-storage==2.19.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-cloud-trace==1.16.1
    # via opentelemetry-exporter-gcp-trace
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.17.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
graphviz==0.20.3
    # via google-adk
greenlet==3.2.2
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
mcp==1.9.2
    # via google-adk
numpy==2.2.6
    # via shapely
opentelemetry-api==1.33.1
    # via
    #   google-adk
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-gcp-trace==1.9.0
    # via google-adk
opentelemetry-resourcedetector-gcp==1.9.0a0
    # via opentelemetry-exporter-gcp-trace
opentelemetry-sdk==1.33.1
    # via
    #   google-adk
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
opentelemetry-semantic-conventions==0.54b1
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
protobuf==5.29.5
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.1.0
    # via
    #   google-adk
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via google-adk
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
rsa==4.9.1
    # via google-auth
shapely==2.1.1
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via google-adk
sse-starlette==2.3.6
    # via mcp
starlette==0.46.2
    # via
    #   fastapi
    #   mcp
typing-extensions==4.13.2
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via tzlocal
tzlocal==5.3.1
    # via google-adk
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via
    #   google-adk
    #   mcp
websockets==15.0.1
    # via google-genai
wrapt==1.17.2
    # via deprecated
zipp==3.22.0
    # via importlib-metadata
